requirements.txt
openai
tiktoken
tenacity
numpy
pandas
requests
python-dotenv
python-telegram-bot==13.15
urllib3==1.26.16
